% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_fn.R
\name{cohort_fn}
\alias{cohort_fn}
\title{Generate a risk data set which is sampled according to a sampling probability dictionary.}
\usage{
cohort_fn(NNN, sampling_probability_dictionary, survival_data, t_star, x_digits,
  t_digits)
}
\arguments{
\item{NNN}{The number of people in the original random sample before
oversampling and undersampling.}

\item{sampling_probability_dictionary}{A named vector,
eg \code{c(A = 1, B = 0.05, C = 1, D = 0.05))}, with names
\code{A}, \code{B}, \code{C}, \code{D},
and values equal to the probability that a person in the corresponding
category is "kept" in the oversampling.  A probability = 1 directs this
function to keep everyone in its corresponding category.}

\item{survival_data}{A logical equal to TRUE if survival data
(e = 0 if censored or dead, e = 1 if disease) rather than competing risk data
(e = 2 if dead) is desired.}

\item{t_star}{A positive real number equal to the right end point of the
duration of the study.}

\item{x_digits}{The number of digits of x_1 and x_2 to round to.}

\item{t_digits}{THe number of digits of t to round to.}
}
\value{
A data frame with columns \code{x_1}, \code{x_2}, \code{category},
\code{sam} (the sampling probability), \code{r_A}, \code{r_B}, \code{t_0},
\code{t_1}, \code{ttt}, \code{eee}.
}
\description{
Obtain a sample that oversamples and undersamples according to categories
defined by two normal covariates x_1 and x_2.
The categories are A, B, C, D determined by
whether or not x_1 and x_2 fall below or above their their respective means.
}
\examples{
NNN = 13
cohort_sampling_probability_dictionary = c(A = 1, B = 0.05, C = 1, D = 0.05)
target_sampling_probability_dictionary = c(A = 1, B = 1,    C = 1, D = 1)
survival_data = FALSE
t_star = 10
x_digits = 2
t_digits = 2
cohort_fn(NNN, cohort_sampling_probability_dictionary, survival_data, t_star, x_digits, t_digits)

}
