% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors}
\alias{colors}
\title{colors}
\usage{
colors(palx = NULL)
}
\arguments{
\item{palx}{Palette name to view the palette colors. Eg. colors("pal_Basic")}
}
\value{
A list of color palettes.
}
\description{
This function loads various color palettes used previously in GCAM
as well as new palettes for Metis modeling to the global environment
}
\details{
List of Color Palettes
\itemize{
\item "pal_rmap"
\item "pal_16"
\item "pal_Basic"
\item "pal_hot"
\item "pal_wet"
\item "pal_div_wet"
\item "pal_div_BlRd"
\item "pal_div_RdBl"
\item "pal_div_BrGn"
\item "pal_div_GnBr"
\item "pal_div_BluRd"
\item "pal_div_RdBlu"
\item "pal_green"
\item "pal_div_BrGn"
\item "pal_div_BlRd"
\item "pal_sankey"
\item "pal_sankey_alpha"
\item "pal_spectral"
\item "pal_ScarcityCat"}
}
\examples{
library(rmap)
# To see all available palettes
names(rmap::colors())
# To use as vector
a <- rmap::colors()$pal_Basic
b <- rmap::colors()[["pal_Basic"]]
}
\keyword{colors,}
\keyword{palette}
