% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_shape.R
\name{df_to_shape}
\alias{df_to_shape}
\title{df_to_shape}
\usage{
df_to_shape(df = NULL)
}
\arguments{
\item{df}{Default = NULL. Fortified ggplot dataframe (e.g. rmap::mapCountriesdf)}
}
\value{
SpatialPolygonDataFrame
}
\description{
Given a fortified ggplot dataframe (e.g. rmap::mapCountriesdf) will choose the
corresponding SpatialPolygonDataFrame map (e.g. rmap::mapCountries)
}
\examples{
library(rmap)
df = rmap::mapCountriesdf
shapex = rmap::df_to_shape(df)
}
\keyword{find}
\keyword{map,}
