\name{df_twoStage}
\alias{df_twoStage}

\title{
df_twoStage
}
\description{
Generate a two-stage sample of people who are subject to competing risk.
In the first stage, generate NTotal people using random sampling (see df_raw).
People with disease (e = 1) are placed into category c = "A", and the remaining people
are placed into category c = "B".  In the second stage, people in category c are kept
with probability ppp[c].  
}
\usage{
df_twoStage(NTotal = 1000, distribution = "lognormal", 
    param1 = -1.8, param2 = 0.4, eta0 = 0.1, eta2 = 0.1, 
    tStar = 10, KKK = 5, ppp = c(A = 1, B = 0.5))
}

\arguments{
  \item{NTotal}{
An integer equal to the number of people in the first stage, generated by a random sample.
}
  \item{distribution}{
A character string equal to "lognormal" or "beta".  This is the distribution used to
generate eta1 which is the rate of the exponential that generates times to disease.
Mean time to disease is equal to 1/eta1.
}
  \item{param1}{
A number that determines the first parameter of the distribution that generates rate 
of disease.  If distribution = "lognormal", log(t1) is normal with mean param1.  If 
distribution = "beta", param1 is the shape1 parameter of rbeta.
}
  \item{param2}{
A number that determines the second parameter of the distribution that generates rate 
of disease.  If distribution = "lognormal", log(t1) is normal with sd param2.  If 
distribution = "beta", param2 is the shape2 parameter of rbeta.
}
  \item{eta0}{
A positive number that determines the rate of censoring.  Times to censoring are generated
according to rexp with rate parameter eta0.
}
  \item{eta2}{
A positive number that determines the rate of death.  Times to death are generated
according to rexp with rate parameter eta2.
}
  \item{tStar}{
A positive number that determines the length of time of the study.  All times are censored at 
tStar.
}
  \item{KKK}{
A positive integer that determines the number of risk groups. 
The program assigns the subjects into approximately equal-sized risk groups taking two-stage
sampling into account.
The risk group assignment is recorded in the column k, 1 for the people with the smallest risk r, 
and so on up to KKK for the people with the largest risk.
}
  \item{ppp}{
A named vector containing two elements ppp["A"] is the resampling probability for category "A", 
and ppp["B"] is the resampling probability for category "B".
}
}
\details{

}
\value{
A list containing three elements
\item{\code{d}}{
A data frame with nrows equall to n["A"] + n["B"] and columns e, t, w, r, c, k.  
These columns are described in help(df_randomSample).
}
\item{\code{N}}{
A named vector of two elements. N["A"], and N["B"] are the number of people in the first stage 
who fell into categories "A" amd "B" respectively. 
}
\item{\code{n}}{
A named vector of two elements. n["A"], and n["B"] are the number of people in the first stage 
who fell into categories "A" amd "B" respectively.
}
}
\references{
For more details and examples, see the package help page at:
http://www.stanford.edu/~ggong/rmap/index.html

}
\author{
David Johnston and Gail Gong

}
\note{

}



\seealso{

}
\examples{
set.seed(1)
df = df_twoStage(1000)
names(df)
head(df$d)
df$N
df$n
}
