% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_gridData_GWPv4To2015}
\alias{example_gridData_GWPv4To2015}
\title{Example grid data}
\format{
tibble
}
\source{
example grid population data https://sedac.ciesin.columbia.edu/data/set/gpw-v4-population-count-rev11/data-download#
}
\usage{
example_gridData_GWPv4To2015
}
\description{
Example grid data
}
\examples{
\dontrun{
 library(rmap);
 gridData <- rmap::example_gridData_GWPv4To2015
}
}
\keyword{datasets}
