% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMapDataClass}
\alias{exampleMapDataClass}
\title{Example data by class to plot in Maps}
\format{
R table or .csv
}
\source{
example data by class
}
\usage{
exampleMapDataClass
}
\description{
Example data by class to plot in Maps
}
\examples{
\dontrun{
 library(rmap);
 mapData <- rmap::exampleMapDataClass
}
}
\keyword{datasets}
