% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMapDataParam}
\alias{exampleMapDataParam}
\title{Example data by param to plot in Maps}
\format{
R table or .csv
}
\source{
example data by Param
}
\usage{
exampleMapDataParam
}
\description{
Example data by param to plot in Maps
}
\examples{
\dontrun{
 library(rmap);
 mapData <- rmap::exampleMapDataParam
}
}
\keyword{datasets}
