% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaHatFn.R
\name{gammaHatFn}
\alias{gammaHatFn}
\title{gammaHat is the proportion of subjects in each risk group}
\usage{
gammaHatFn(baseArgs)
}
\arguments{
\item{baseArgs}{A list provided by \code{baseArgsFn}.
The objects required here are
\code{c}
\code{K}, \code{k}, and
\code{weight}.}
}
\value{
A numeric vector with one element for each risk group.  
The \code{k}-th element of this vector is the proportion of 
subjects in risk group k.  
See equation (45) in "rmap-formulas-v02.pdf" from the website.
}
\description{
This function computes gammaHat, the proportion of subjects in each
risk group.  This function is called internally by 
grouped \code{rmap}.
It takes two-stage sampling weights into account if a two-stage 
sampling design was used.
}
\examples{
set.seed(1)
d = df_twoStage()
e = d$d$e
t = d$d$t
r = d$d$r
tStar = 10
design = list(c = d$d$c, N_two_stage = d$N)
riskGroup = list(K = 2)
rSummary = "median"
bootstrap = FALSE
ba = baseArgsFn(e, t, r, tStar, 
  design, riskGroup, rSummary, bootstrap)
gh1 = gammaHatFn(ba)
gh1

set.seed(1)
d = df_randomSample()
baseArgs = baseArgsFn(e, t, r, tStar,
  design, riskGroup, rSummary, bootstrap)
gh2 = gammaHatFn(baseArgs)
gh2

}
