% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_bootstrap_sds_fn.R
\name{grouped_bootstrap_sds_fn}
\alias{grouped_bootstrap_sds_fn}
\title{The bootstrap part of \code{rmap_fn}}
\usage{
grouped_bootstrap_sds_fn(baseArgs)
}
\arguments{
\item{baseArgs}{A list provided by \code{baseArgsFn}.  The objects
required by \code{grouped_bootstrap_sds_fn} are 
\code{c}
\code{e},
\code{K}, \code{k},
\code{N_bootstraps},
\code{N_nonzero_events},
\code{N_two_stage}, \code{n_two_stage},
\code{r},
\code{rSummary},
\code{t},
\code{target_category},
\code{tStar}, and
\code{weight}}
}
\value{
A list containing
\itemize{
\item{\code{grouped_bootstrap_sds}: }{A data.frame containing \code{K} rows,
one for each risk group, and the following columns:
\itemize{
\item{\code{sd_boot}: }{The bootstrap estimate of teh
standard deviation of the \code{pi_hat}.
\code{pi_hat}.
}
\item{\code{lower}: }{The lower limit of the confidence interval for
the outcome probability.
}
\item{\code{upper}: }{The upper limit of the confidence interval for
the outcome probability.
}
\item{\code{in_ci_boot}: }{A character "yes" or "no" indicating if the summary
provided by \code{baseArgs$rSummary} falls in the confidence interval.
}
}
}
\item{\code{concordance_ci}: }{A named vector containing \code{lower} 
and \code{upper} the bounds of a confidence interval for the concordance. 
}
}
}
\description{
For each risk group, 
calculate the bootstrap variance of the outcome probability estimate
\code{pi_hat} and produce a confidence interval for the
outcome probability.
}
\examples{
set.seed(1)
twoStageSample = df_twoStage(60)
tStar = 10
xxx = twoStageSample$d
e = xxx$e
t = round(xxx$t, 1)
r = xxx$r
N = twoStageSample$N
design = list(N_two_stage = N, c = xxx$c)
riskGroup = list(K = 2)
rSummary = "mean"
bootstrap = 100
baseArgs = baseArgsFn(e, t, r, tStar, design, riskGroup, rSummary, bootstrap)
grouped_bootstrap_sds_fn(baseArgs)

}
