% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_one_bootstrap_fn.R
\name{interpolate_one_bootstrap_fn}
\alias{interpolate_one_bootstrap_fn}
\title{interpolate_one_bootstrap_fn}
\usage{
interpolate_one_bootstrap_fn(one_bootstrap, rho_uni)
}
\arguments{
\item{one_bootstrap}{A numerical vector of the same
length as \code{rho_uni}.  This vector corresponds
to one bootstrap replication of \code{pi_hat_nn_fn}.
If \code{rho} in \code{rho_uni} appears in the bootstrap
\code{one_bootstrap} is equal to \code{pi_hat} at this
\code{rho}; if \code{rho} is missing, \code{one_bootstrap}
is equal to \code{NA} at this \code{rho}.}

\item{rho_uni}{A numerical vector containt the ordered 
unique values of assigned risk of the original data.}
}
\value{
A numerical vector of the same lenght as 
\code{rho_uni}. The same vector as \code{one_bootstrap}
but with all the \code{NA}s replaced with
interpolated values.
}
\description{
Interpolate those \code{rho} missing from a bootstrap
}
\examples{
options(digits = 3)
set.seed(1)
xxx = df_randomSample(10)
e = xxx$e
t = xxx$t
r = round(xxx$r, 2)
tStar = 10
design = "randomSample"
epsilon = nrow(xxx)^(-1/3)
riskGroup = list(epsilon = epsilon)
rSummary = "mean"
bootstrap = 3
confidenceLevel = 0.95
multicore = FALSE
verbose = FALSE
baseArgs = baseArgsFn(e, t, r, tStar, design, riskGroup, rSummary, bootstrap,
                      confidenceLevel, multicore, verbose)
e = baseArgs$e
t = baseArgs$t
tStar = baseArgs$tStar
baseArgs$e = ifelse(t > tStar, 0, e)
baseArgs$t = ifelse(t > tStar, tStar, t)
estimate = pi_hat_nn_fn(baseArgs)
estimate
bootstraps_raw = lapply(1:baseArgs$N_bootstraps, function(n_bootstrap){
  base_args_boot = base_args_boot_fn(baseArgs)
  pi_hat_nn_fn(base_args_boot)
})
bootstraps_raw
gathered = gather_fn(bootstraps_raw, estimate[, "rho"])
gathered
interpolated = t(apply(gathered, 1, interpolate_one_bootstrap_fn, estimate[, "rho"]))
interpolated

}
