% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaHatFn.R
\name{lambdaHatFn}
\alias{lambdaHatFn}
\title{The hazard for each risk group}
\usage{
lambdaHatFn(baseArgs)
}
\arguments{
\item{baseArgs}{A list provided by \code{baseArgsFn}.
Objects required by \code{lambdaHatFn} are
\code{c},
\code{e},
\code{K},
\code{k},
\code{t},
\code{tStar}, and
\code{weight}.}
}
\value{
A list with \code{K} lists, each list containing
\itemize{
\item{\code{tau}: }{
Ordered failure times for the risk group where failure time
can be times for which \code{e = 1} or {2}.
}
\item{\code{NAR}: }{Number at risk at time \code{tau}.  
A logical matrix with one row for each subject in the risk
group and one column for each \code{tau}.  
Indexing the subjects in the risk group with \code{n}
and the elements of \code{tau} with \code{m}, 
\code{NAR[n, m]} is \code{TRUE} if \code{t[n] >= tau[m]}.
See equation (29) in "rmap-formulas-v02.pdf" from the website.
}
\item{\code{DDD}: }{Disease or death at times \code{tau}.
\code{DDD} is a list of two elements, the first for 
\code{e = 1} (outcome) and the second for \code{e = 2} (death).
Both elements is a logical matrix with one row for each
subject in the risk group and one column for each \code{tau}.
The \code{(n, m)}-th element of  \code{DDD[[1]]} is \code{TRUE}
if \code{e[n] == 1} and \code{t[n] == tau[m]}.
The \code{(n, m)}-th element of \code{DDD[[2]]} is \code{TRUE}
if \code{e[n] == 2} and \code{t[n] == tau[m]}.
See equation (30) in "rmap-formulas-v02.pdf" from the website.
}
\item{\code{denom}: }{The number at risk at time \code{tau}
calculated by getting the column-wise weighted sums of \code{NAR}.
See equation(61)
}
\item{\code{lambdaHat}: }{The estimated hazard rate at times 
\code{tau} for event = 1 (disease) or event = 2 (death from 
other causes).
}
}
}
\description{
Compute \code{lambdaHat}, \code{NAR}, \code{tau}, and \code{denom}
for each risk group.
}
\examples{
set.seed(1)
twoStageSample = df_twoStage(20)
tStar = 10
xxx = twoStageSample$d
e = xxx$e
t = round(xxx$t, 1)
r = xxx$r
N = twoStageSample$N
design = list(N_two_stage = N, c = xxx$c)
riskGroup = list(K = 2)
rSummary = "mean"
bootstrap = FALSE
baseArgs = baseArgsFn(e, t, r, tStar, design, riskGroup, rSummary, bootstrap)
with(baseArgs, data.frame(e, t, r, k, weight))
lambdaHat = lambdaHatFn(baseArgs)
names(lambdaHat)
lambdaHat[["k1"]]
lambdaHat[["k2"]]

}
