% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{map}
\usage{
map(
  data = NULL,
  fillColumn = NULL,
  fileName = "map",
  save = T,
  theme = NULL,
  folder = paste(getwd(), "/outputs", sep = ""),
  labels = F,
  labelForce = 0,
  labelColor = "black",
  labelSize = 3,
  labelAlpha = 1,
  labelFill = NA,
  labelBorderSize = NA,
  shapeFolder = NULL,
  shapeFile = NULL,
  subRegCol = "subRegion",
  nameAppend = "",
  legendTitle = NULL,
  legendType = "kmeans",
  legendBreaksn = 5,
  legendFixedBreaks = NULL,
  animateOn = T,
  fps = 1,
  cropToBoundary = T,
  underLayer = NULL,
  underLayerColor = "gray40",
  underLayerFill = "gray90",
  underLayerLwd = 0.1,
  underLayerAlpha = 1,
  underLayerLabels = F,
  overLayerLabels = F,
  overLayer = NULL,
  overLayerColor = "gray40",
  overLayerFill = NA,
  overLayerLwd = 0.2,
  overLayerAlpha = 0,
  zoom = -1,
  zoomx = NULL,
  zoomy = NULL,
  asp = 1.2,
  projX = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  width = 7,
  height = 8,
  scenRef = NULL,
  scenDiff = NULL,
  scaleRange = NULL,
  scaleRangeDiffAbs = NULL,
  scaleRangeDiffPrcnt = NULL,
  xRef = NULL,
  xDiff = NULL,
  scaleRangeDiffxAbs = NULL,
  scaleRangeDiffxPrcnt = NULL,
  facetCols = NULL,
  facetRows = NULL,
  mapTitle = NULL,
  numeric2Cat_list = NULL,
  pdfpng = "png",
  legendDigitsOverride = NULL,
  legendSingleColor = "white",
  legendSingleValue = NULL,
  classPalette = NULL,
  classPaletteDiff = "pal_div_BluRd",
  colorNA = "gray50",
  showNA = F,
  ncol = 3,
  size = 12,
  alpha = 1,
  background = F,
  transparent = T,
  legendShow = T
)
}
\arguments{
\item{data}{Default = NULL,}

\item{fillColumn}{(Optional). Default = NULL. Only for direct map plotting.}

\item{fileName}{(Optional). Default = "map". Only for direct map plotting.}

\item{save}{(Optional). Default = T. Only for direct map plotting.}

\item{theme}{Default = NULL,}

\item{folder}{Default = paste(getwd(),"/outputs",sep=""),}

\item{labels}{Default = F,}

\item{labelForce}{Default = 0,}

\item{labelColor}{Default = "black",}

\item{labelSize}{Default = 3}

\item{labelAlpha}{Default = 0.7,}

\item{labelFill}{Default = NA,}

\item{labelBorderSize}{Default = NA}

\item{shapeFolder}{Default = paste(getwd(),"/dataFiles/gis/admin_gadm36",sep=""),}

\item{shapeFile}{Default = paste("gadm36_1",sep=""),}

\item{subRegCol}{Default ="subRegion",}

\item{nameAppend}{Default =""}

\item{legendTitle}{Default = NULL}

\item{legendType}{Default ="kmeans", Options include c("pretty","kmeans","freescale","all")}

\item{legendBreaksn}{Default = "5",}

\item{legendFixedBreaks}{Default = NULL,}

\item{animateOn}{Default = T,}

\item{fps}{Default = 1,}

\item{cropToBoundary}{Default = F}

\item{underLayer}{Default = NULL}

\item{underLayerColor}{Default = "gray40"}

\item{underLayerFill}{Default = "gray90"}

\item{underLayerLwd}{Default = 0.5}

\item{underLayerAlpha}{Default = 1}

\item{underLayerLabels}{Default = F}

\item{overLayerLabels}{Default = F}

\item{overLayer}{Default = NULL}

\item{overLayerColor}{Default = "gray40"}

\item{overLayerFill}{Default = NA}

\item{overLayerLwd}{Default = 0.5}

\item{overLayerAlpha}{Default = 0}

\item{zoom}{Default =-1. Zoom into or out of map. Positive values zoom in and negative out.}

\item{zoomx}{Default = NULL. Zoom into or out of map along x. Positive values zoom in and negative out.}

\item{zoomy}{Default = NULL. Zoom into or out of map along y. Positive values zoom in and negative out.}

\item{asp}{Default = 1.2. Aspect ratio of lat and lon.}

\item{projX}{Default = projX="+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"}

\item{width}{Default =9}

\item{height}{Default =7}

\item{scenRef}{Reference Scenario. Default = NULL}

\item{scenDiff}{Scenarios to Diff. Default = NULL}

\item{scaleRange}{Default NULL. A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min}

\item{scaleRangeDiffAbs}{Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min}

\item{scaleRangeDiffPrcnt}{Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min}

\item{xRef}{Reference year. Default = NULL}

\item{xDiff}{years to Diff. Default = NULL}

\item{scaleRangeDiffxAbs}{Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min}

\item{scaleRangeDiffxPrcnt}{Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min}

\item{facetCols}{Default ="multiFacetRow",}

\item{facetRows}{Default ="multiFacetCol",}

\item{mapTitle}{Default=NULL}

\item{numeric2Cat_list}{Default=NULL,}

\item{pdfpng}{Save IO figures as pdf or png. Type=String. Options: 'pdf' or 'png'. Default = 'png'}

\item{legendDigitsOverride}{Default=NULL}

\item{legendSingleColor}{Default="white"}

\item{legendSingleValue}{Default =NULL,}

\item{classPalette}{Default = NULL}

\item{classPaletteDiff}{Default = "pal_div_BrGn"}

\item{colorNA}{Default = "gray50"}

\item{showNA}{Default = T}

\item{ncol}{Default = 3. Number of columns to wrap maps}

\item{size}{Default = 12. Text size of plots.}

\item{alpha}{Default = 1. Transparency of fill colors.}

\item{background}{Default = F. Add background water color, border and default underlayer map.}

\item{transparent}{Default = T. To make map background transparent for maps without backgrounds.}

\item{legendShow}{Default = T}
}
\value{
Returns the formatted data used to produce chart

A list of maps
}
\description{
This function produce different kinds of maps for the rmap package.
Each figure is accompanied with a csv table.
}
\keyword{charts,}
\keyword{diffplots}
