% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_find.R
\name{map_find}
\alias{map_find}
\title{map_find}
\usage{
map_find(
  data,
  subRegion = "subRegion",
  value = "value",
  x = NULL,
  class = "class",
  scenario = "scenario"
)
}
\arguments{
\item{data}{data table to be processed}

\item{subRegion}{Default ="subRegion",}

\item{value}{Default = "value",}

\item{x}{Default = NULL}

\item{class}{Default = "class"}

\item{scenario}{Default = "scenario"}
}
\value{
dataframe with modified subRegions, subRegion shapefile and subRegion type
}
\description{
Given a data.frame with a subRegion column, this function searches for an appropriate map from the pre-loaded rmap maps.
}
\keyword{find}
\keyword{map,}
