% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapCountries}
\alias{mapCountries}
\title{World Map of Countries}
\format{
sf
}
\source{
Made with Natural Earth. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
mapCountries
}
\description{
World Map of Countries
}
\examples{
\dontrun{
 library(rmap)
 plot(mapCountries[,"subRegion"])
 head(mapCountries)
}
}
\keyword{datasets}
