% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapCountriesdf}
\alias{mapCountriesdf}
\title{World Map of Countries}
\format{
A SpatialPolygonsDataFrame
}
\source{
Made with Natural Earth. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
mapCountriesdf
}
\description{
World Map of Countries
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapCountriesdf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapCountriesdf)
}
}
\keyword{datasets}
