% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapCountriesUS52}
\alias{mapCountriesUS52}
\title{Merge of Countries and US52}
\format{
sf
}
\source{
JGCRI
}
\usage{
mapCountriesUS52
}
\description{
Merge of Countries and US52
}
\examples{
\dontrun{
library(rmap)
plot(mapCountriesUS52[,"subRegion"])
head(mapCountriesUS52)
}
}
\keyword{datasets}
