% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapCountriesUS52df}
\alias{mapCountriesUS52df}
\title{Merge of Countries and US52}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapCountriesUS52df
}
\description{
Merge of Countries and US52
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapCountriesUS52df, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapCountriesUS52df)
}
}
\keyword{datasets}
