% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFind.R
\name{mapFind}
\alias{mapFind}
\title{mapFind}
\usage{
mapFind(dataTbl)
}
\arguments{
\item{dataTbl}{Palette name to view the palette colors. Eg. colors("pal_Basic")}
}
\value{
dataframe with modified subRegions, subRegion shapefile and subRegion type
}
\description{
Given a data.frame with a subRegion column, this function searches for an appropriate map from the pre-loaded rmap maps.
}
\examples{
library(rmap)
data = data.frame(subRegion=c("FL","ID","MO"),value=c(-2,3,14))
rmap::mapFind(data)
}
\keyword{find}
\keyword{map,}
