% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasinsdf}
\alias{mapGCAMBasinsdf}
\title{GCAM Basins}
\format{
A SpatialPolygonsDataFrame
}
\source{
Modified versions of shapefiles from \url{https://zenodo.org/record/4688451#.YdMNTmjMJPY}
}
\usage{
mapGCAMBasinsdf
}
\description{
GCAM Basins
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapGCAMBasinsdf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapGCAMBasinsdf)
}
}
\keyword{datasets}
