% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasinsUS49}
\alias{mapGCAMBasinsUS49}
\title{Cropped of GCAM Basins and US49}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapGCAMBasinsUS49
}
\description{
Cropped of GCAM Basins and US49
}
\examples{
\dontrun{
library(rmap)
head(mapGCAMBasinsUS49)
}
}
\keyword{datasets}
