% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasinsUS49df}
\alias{mapGCAMBasinsUS49df}
\title{Cropped of GCAM Basins and US49}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapGCAMBasinsUS49df
}
\description{
Cropped of GCAM Basins and US49
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapGCAMBasinsUS49df, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapGCAMBasinsUS49df)
}
}
\keyword{datasets}
