% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasinsUS52}
\alias{mapGCAMBasinsUS52}
\title{Cropped map of GCAM Basins and US52}
\format{
sf
}
\source{
JGCRI
}
\usage{
mapGCAMBasinsUS52
}
\description{
Cropped map of GCAM Basins and US52
}
\examples{
\dontrun{
library(rmap)
plot(mapGCAMBasinsUS52[,"subRegion"])
head(mapGCAMBasinsUS52)
}
}
\keyword{datasets}
