% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasinsUS52df}
\alias{mapGCAMBasinsUS52df}
\title{Cropped of GCAM Basins and US52}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapGCAMBasinsUS52df
}
\description{
Cropped of GCAM Basins and US52
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapGCAMBasinsUS52df, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapGCAMBasinsUS52df)
}
}
\keyword{datasets}
