% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMLand}
\alias{mapGCAMLand}
\title{GCAM Land}
\format{
sf
}
\source{
Modified versions of shapefiles from \url{https://zenodo.org/record/4688451#.YdMNTmjMJPY}
}
\usage{
mapGCAMLand
}
\description{
Intersection of mapGCAMBasins and mapGCAMReg32
}
\examples{
\dontrun{
 library(rmap)
 plot(mapGCAMLand[,"subRegion"])
 head(mapGCAMLand)
}
}
\keyword{datasets}
