% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMLanddf}
\alias{mapGCAMLanddf}
\title{GCAM Land}
\format{
A SpatialPolygonsDataFrame
}
\source{
From the moirai project.
Shapefile received directly from Alan V. Di Vittorio.
Component files available at:
\url{https://github.com/JGCRI/moirai/blob/master/ancillary/moirai_valid_boundaries.zip}
\url{https://github.com/JGCRI/moirai/blob/build-r-package/tests/testthat/test_compare_raster_w_outputs.R}
}
\usage{
mapGCAMLanddf
}
\description{
GCAM Land
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapGCAMLanddf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapGCAMLanddf)
}
}
\keyword{datasets}
