% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMLandUS49df}
\alias{mapGCAMLandUS49df}
\title{GCAM Land cropped to US49}
\format{
A SpatialPolygonsDataFrame
}
\source{
Created from rmap maps.
}
\usage{
mapGCAMLandUS49df
}
\description{
GCAM Land cropped to US49
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapGCAMLandUS49df, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapGCAMLandUS49df)
}
}
\keyword{datasets}
