% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMLandUS52df}
\alias{mapGCAMLandUS52df}
\title{GCAM Land cropped to US52}
\format{
A SpatialPolygonsDataFrame
}
\source{
Created from rmap maps
}
\usage{
mapGCAMLandUS52df
}
\description{
GCAM Land cropped to US52
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapGCAMLandUS52df, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapGCAMLandUS52df)
}
}
\keyword{datasets}
