% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMReg32}
\alias{mapGCAMReg32}
\title{GCAM 32 Regions}
\format{
sf
}
\source{
Modified versions of shapefiles from \url{https://zenodo.org/record/4688451#.YdMNTmjMJPY}
}
\usage{
mapGCAMReg32
}
\description{
GCAM 32 Regions
}
\examples{
\dontrun{
 library(rmap)
 plot(mapGCAMReg32[,"subRegion"])
 head(mapGCAMReg32)
}
}
\keyword{datasets}
