% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMReg32EU}
\alias{mapGCAMReg32EU}
\title{GCAM 32 Regions with EU countries}
\format{
sf
}
\source{
Modified versions of shapefiles from \url{https://zenodo.org/record/4688451#.YdMNTmjMJPY}
}
\usage{
mapGCAMReg32EU
}
\description{
GCAM 32 Regions with EU countries
}
\examples{
\dontrun{
 library(rmap)
 plot(mapGCAMReg32EU[,"subRegion"])
 head(mapGCAMReg32EU)
}
}
\keyword{datasets}
