% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMReg32Uruguay}
\alias{mapGCAMReg32Uruguay}
\title{GCAM 32 Regions with Uruguay}
\format{
A SpatialPolygonsDataFrame
}
\source{
Modified versions of shapefiles from \url{https://zenodo.org/record/4688451#.YdMNTmjMJPY}
}
\usage{
mapGCAMReg32Uruguay
}
\description{
GCAM 32 Regions with Uruguay
}
\examples{
\dontrun{
 library(sp); library(rmap)
 sp::plot(mapGCAMReg32Uruguay)
 head(mapGCAMReg32Uruguay@data)
}
}
\keyword{datasets}
