% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMReg32US52df}
\alias{mapGCAMReg32US52df}
\title{Merge of GCAM 32 and US52}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapGCAMReg32US52df
}
\description{
Merge of GCAM 32 and US52
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapGCAMReg32US52df, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapGCAMReg32US52df)
}
}
\keyword{datasets}
