% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapHydroShed3}
\alias{mapHydroShed3}
\title{HydroSHEDS level 3}
\format{
sf
}
\source{
Lehner, B., Grill G. (2013): Global river hydrography and network routing:
baseline data and new approaches to study the world’s large river systems.
Hydrological Processes, 27(15): 2171–2186. \url{https://www.hydrosheds.org/page/hydrobasins}
}
\usage{
mapHydroShed3
}
\description{
HydroSHEDS level 3
}
\examples{
\dontrun{
 library(rmap)
 plot(mapHydroShed3[,"subRegion"])
 head(mapHydroShed3)
}
}
\keyword{datasets}
