% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapHydroShed3df}
\alias{mapHydroShed3df}
\title{HydroSHEDS level 3}
\format{
A SpatialPolygonsDataFrame
}
\source{
Lehner, B., Grill G. (2013): Global river hydrography and network routing:
baseline data and new approaches to study the world’s large river systems.
Hydrological Processes, 27(15): 2171–2186. \url{https://www.hydrosheds.org/page/hydrobasins}
}
\usage{
mapHydroShed3df
}
\description{
HydroSHEDS level 3
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapHydroShed3df, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapHydroShed3df)
}
}
\keyword{datasets}
