% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapIntersectGCAMBasin32RegUruguaydf}
\alias{mapIntersectGCAMBasin32RegUruguaydf}
\title{Intersection of GCAM Basins and GCAM 32 Regions + Uruguay.}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapIntersectGCAMBasin32RegUruguaydf
}
\description{
Intersection of GCAM Basins and GCAM 32 Regions + Uruguay.
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapIntersectGCAMBasin32RegUruguaydf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapIntersectGCAMBasin32RegUruguaydf)
}
}
\keyword{datasets}
