% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapIntersectGCAMBasinUS52County}
\alias{mapIntersectGCAMBasinUS52County}
\title{Intersection of GCAM Basins and US 52 County Regions.}
\format{
sf
}
\source{
JGCRI
}
\usage{
mapIntersectGCAMBasinUS52County
}
\description{
Intersection of GCAM Basins and US 52 County Regions.
}
\examples{
\dontrun{
library(rmap)
plot(mapInteresectGCAMBasinUS52County[,"subRegion"])
head(mapIntersectGCAMBasinUS52County)
}
}
\keyword{datasets}
