% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapIntersectGCAMBasinUS52Countydf}
\alias{mapIntersectGCAMBasinUS52Countydf}
\title{Intersection of GCAM Basins and US 52 County Regions.}
\format{
R tibble
}
\source{
JGCRI
}
\usage{
mapIntersectGCAMBasinUS52Countydf
}
\description{
Intersection of GCAM Basins and US 52 County Regions.
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapIntersectGCAMBasinUS52Countydf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapIntersectGCAMBasinUS52Countydf)
}
}
\keyword{datasets}
