% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_country_gcam32}
\alias{mapping_country_gcam32}
\title{mapping_country_gcam32}
\format{
R table
}
\source{
rmap
}
\usage{
mapping_country_gcam32
}
\description{
mapping_country_gcam32
}
\examples{
\dontrun{
 library(rmap);
 mapping_country_gcam32 <- rmap::mapping_country_gcam32
}
}
\keyword{datasets}
