% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_gcambasins}
\alias{mapping_gcambasins}
\title{mapping_gcambasins}
\format{
R table
}
\source{
rmap
}
\usage{
mapping_gcambasins
}
\description{
mapping_gcambasins
}
\examples{
\dontrun{
 library(rmap);
 mapping_gcambasins <- rmap::mapping_gcambasins
}
}
\keyword{datasets}
