% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_US52}
\alias{mapping_US52}
\title{mapping_US52}
\format{
R table
}
\source{
rmap
}
\usage{
mapping_US52
}
\description{
mapping_US52
}
\examples{
\dontrun{
 library(rmap);
 mapping_US52 <- rmap::mapping_US52
}
}
\keyword{datasets}
