% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapplot.R
\name{mapplot}
\alias{mapplot}
\title{map}
\usage{
mapplot(
  dataPolygon = NULL,
  fillColumn = NULL,
  dataGrid = NULL,
  shpFolder = NULL,
  shpFile = NULL,
  theme = NULL,
  fillPalette = "Spectral",
  labels = F,
  width = 9,
  height = 7,
  legendShow = F,
  legendTitle = NULL,
  legendBreaksn = 5,
  legendBreaks = NULL,
  pdfpng = "png",
  underLayer = NULL,
  underLayerColor = "gray40",
  underLayerFill = "gray40",
  underLayerLwd = 0.5,
  underLayerAlpha = 1,
  overLayer = NULL,
  overLayerColor = "gray40",
  overLayerFill = NA,
  overLayerLwd = 0.5,
  overLayerAlpha = 0,
  zoom = -1,
  zoomx = NULL,
  zoomy = NULL,
  printFig = T,
  fileName = "map",
  folder = paste(getwd(), "/outputs", sep = ""),
  facetRows = NULL,
  facetCols = NULL,
  mapTitle = NULL,
  numeric2Cat_list = NULL,
  legendDigits = NULL,
  legendDigitsOverride = NULL,
  legendSingleColor = "white",
  legendSingleValue = NULL,
  colorNA = "gray50",
  showNA = F,
  ncol = 3,
  size = 12,
  alpha = 1,
  background = F
)
}
\arguments{
\item{dataPolygon}{Default = NULL}

\item{fillColumn}{Default = NULL # Or give column with data}

\item{dataGrid}{Default = NULL}

\item{shpFolder}{Default = paste(getwd(),"/dataFiles/gis/admin_gadm36_1",sep Default = "")}

\item{shpFile}{Default = paste("gadm36_1",sep Default = "")}

\item{theme}{Default = NULL}

\item{fillPalette}{Default = "Spectral"}

\item{labels}{Default = FALSE}

\item{width}{Default = 9}

\item{height}{Default = 7}

\item{legendShow}{Default = F}

\item{legendTitle}{Default = NULL}

\item{legendBreaksn}{Default = "5"}

\item{legendBreaks}{Default = NULL}

\item{pdfpng}{Default = "png"}

\item{underLayer}{Default = NULL}

\item{underLayerColor}{Default = "gray40"}

\item{underLayerFill}{Default = "gray40"}

\item{underLayerLwd}{Default = 0.5}

\item{underLayerAlpha}{Default = 1}

\item{overLayer}{Default = NULL}

\item{overLayerColor}{Default = "gray40"}

\item{overLayerFill}{Default = NA}

\item{overLayerLwd}{Default = 0.5}

\item{overLayerAlpha}{Default = 0}

\item{zoom}{Default =-1. Zoom into or out of map. Positive values zoom in and negative out.}

\item{zoomx}{Default = NULL. Zoom into or out of map along x. Positive values zoom in and negative out.}

\item{zoomy}{Default = NULL. Zoom into or out of map along y. Positive values zoom in and negative out.}

\item{printFig}{Default = T}

\item{fileName}{Default = "map"}

\item{folder}{Default = paste(getwd(),"/outputs",sep Default = "")}

\item{facetRows}{Default=NULL}

\item{facetCols}{Default=NULL}

\item{mapTitle}{Default=NULL}

\item{numeric2Cat_list}{Default=NULL}

\item{legendDigits}{Default = NULL}

\item{legendDigitsOverride}{Default=NULL}

\item{legendSingleColor}{Default="white"}

\item{legendSingleValue}{Default=NULL}

\item{colorNA}{Default = "gray50"}

\item{showNA}{Default = T}

\item{ncol}{Default = 3.  Number of columns to wrap maps}

\item{size}{Default = 12. Text size of plots.}

\item{alpha}{Default = 1. Transparency of fill colors.}

\item{background}{Default = F. Add background water color, border and default underlayer map.}
}
\value{
Returns the formatted data used to produce chart
}
\description{
This function produce different kinds of maps for the rmap package.
Each figure is accompanied with a csv table.
}
\keyword{charts,}
\keyword{diffplots}
