% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapsSubRegions}
\alias{mapsSubRegions}
\title{List of subRegions in Pre-loaded maps}
\format{
List
}
\source{
Compiled from each map
}
\usage{
mapsSubRegions
}
\description{
List of subRegions in Pre-loaded maps
}
\examples{
\dontrun{
 library(rmap);
 names(mapsSubRegions)
}
}
\keyword{datasets}
