% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapStates}
\alias{mapStates}
\title{World Map of States}
\format{
sf
}
\source{
Made with Natural Earth. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
mapStates
}
\description{
World Map of States
}
\examples{
\dontrun{
 library(rmap)
 plot(mapStates[,"subRegion"])
 head(mapStates)
}
}
\keyword{datasets}
