% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapStatesdf}
\alias{mapStatesdf}
\title{World Map of States}
\format{
A SpatialPolygonsDataFrame
}
\source{
Made with Natural Earth. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
mapStatesdf
}
\description{
World Map of States
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapStatesdf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapStatesdf)
}
}
\keyword{datasets}
