% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS49Countydf}
\alias{mapUS49Countydf}
\title{US 49 States
Excludes Alaska, Hawaii and Puerto Rico. Includes DC.}
\format{
A SpatialPolygonsDataFrame
}
\source{
Made with Natural Earth. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
mapUS49Countydf
}
\description{
US 49 States
Excludes Alaska, Hawaii and Puerto Rico. Includes DC.
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapUS49Countydf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapUS49Countydf)
}
}
\keyword{datasets}
