% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS49HUC2df}
\alias{mapUS49HUC2df}
\title{USGS Hydrological Unit Code (HUC)}
\format{
A SpatialPolygonsDataFrame
}
\source{
\url{https://water.usgs.gov/GIS/huc.html} \url{https://nrcs.app.box.com/v/huc}
}
\usage{
mapUS49HUC2df
}
\description{
USGS Hydrological Unit Code (HUC)
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapUS49HUC2df, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapUS49HUC2df)
}
}
\keyword{datasets}
