% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS52Compactdf}
\alias{mapUS52Compactdf}
\title{US 52 States Compact
Includes Alaska, Hawaii and Puerto Rico as well as DC (Re-positioned)}
\format{
A SpatialPolygonsDataFrame
}
\source{
US Census bureau. \url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
mapUS52Compactdf
}
\description{
US 52 States Compact
Includes Alaska, Hawaii and Puerto Rico as well as DC (Re-positioned)
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapUS52Compactdf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapUS52Compactdf)
}
}
\keyword{datasets}
