% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS52County}
\alias{mapUS52County}
\title{US 52 Counties
Includes Alaska, Hawaii and Puerto Rico as well as DC.}
\format{
sf
}
\source{
US Census bureau. \url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
mapUS52County
}
\description{
US 52 Counties
Includes Alaska, Hawaii and Puerto Rico as well as DC.
}
\examples{
\dontrun{
 library(rmap)
 plot(mapUS52County[,"subRegion"])
 head(mapUS52County)
}
}
\keyword{datasets}
