% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS52CountyCompact}
\alias{mapUS52CountyCompact}
\title{US 52 Counties Compact
Includes Alaska, Hawaii and Puerto Rico as well as DC (Repositioned)}
\format{
sf
}
\source{
US Census bureau. \url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
mapUS52CountyCompact
}
\description{
US 52 Counties Compact
Includes Alaska, Hawaii and Puerto Rico as well as DC (Repositioned)
}
\examples{
\dontrun{
 library(rmap)
 plot(mapUS52CountyCompact[,"subRegion"])
 head(mapUS52CountyCompact)
}
}
\keyword{datasets}
