% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS52HUC4}
\alias{mapUS52HUC4}
\title{USGS Hydrological Unit Code (HUC)}
\format{
sf
}
\source{
\url{https://water.usgs.gov/GIS/huc.html} \url{https://nrcs.app.box.com/v/huc}
}
\usage{
mapUS52HUC4
}
\description{
USGS Hydrological Unit Code (HUC)
}
\examples{
\dontrun{
 library(rmap)
 plot(mapUS52HUC4[,"subRegion"])
 head(mapUS52HUC4)
}
}
\keyword{datasets}
