% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapr.R
\docType{data}
\name{N_first_stage}
\alias{N_first_stage}
\title{The data from the first stage}
\format{A named vector with two elements names \code{"A"} and \code{"B"}}
\usage{
data(N_first_stage)
}
\description{
A named vector that counts the number of people
in the first stage in each two-stage category.
The second stage data are in 
 \code{two_stage_sample_example}
}
\keyword{rmap_datasets}
