\name{nlapply}
\alias{nlapply}

\title{
Blah
}
\description{
describe
}
\usage{
nlapply(X, FUN, ...)
}

\arguments{
  \item{X}{
a named vector or list
}
  \item{FUN}{
A function
}
  \item{\dots}{
dots
}
}
\details{

}
\value{
A named list
}
\references{

}
\author{
GG
}
\note{

}



\seealso{

}
\examples{
larry = c(a = 14, b = 42)
nlapply(larry, function(this) this^2)
}


