% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_hat_nn_fn.R
\name{pi_hat_nn_fn}
\alias{pi_hat_nn_fn}
\title{A pi_hat for each distinct assigned risk}
\usage{
pi_hat_nn_fn(baseArgs)
}
\arguments{
\item{baseArgs}{A list provided by \code{baseArgsFn}.
The objects required to run \code{pi_hat_nn_fn} are
\code{c},
\code{e},
\code{epsilon},
\code{r}
\code{sampling},
\code{t}
\code{verbose}, and
\code{weight}.}
}
\value{
A data.frame containing the columns
\code{rho} an ordered vector of distinct assigned risks
and \code{pi_hat} the corresponding
epsilon kernel nearest neighbor
estimate of outcome probability.
}
\description{
An epsilon kernel nearest neighbor estimate of outcome
probabilty for each distinct value of assigned risk
}
\examples{
set.seed(1)
xxx = df_randomSample(40)
e = xxx$e
t = xxx$t
r = round(xxx$r, 3)
tStar = 10
design = "randomSample"
epsilon = nrow(xxx)^(-1/3)
riskGroup = list(epsilon = epsilon)
rSummary = "mean"
bootstrap = 20
confidenceLevel = 0.95
multicore = FALSE
verbose = TRUE
baseArgs = baseArgsFn(e, t, r, tStar, design, riskGroup, rSummary, bootstrap,
                      confidenceLevel, multicore, verbose)
e = baseArgs$e
t = baseArgs$t
tStar = baseArgs$tStar
baseArgs$e = ifelse(t > tStar, 0, e)
baseArgs$t = ifelse(t > tStar, tStar, t)
estimate = pi_hat_nn_fn(baseArgs)
estimate
ggplot(estimate, aes(x = rho, y = pi_hat)) + geom_line() 
}
