% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_individual_risk_plot.R
\name{pretty_individual_risk_plot}
\alias{pretty_individual_risk_plot}
\title{An annotated indivualized attribute diagram plot using results from 
\code{rmap_individual}}
\usage{
pretty_individual_risk_plot(individual,
  title = "Individualized Attribute Diagram", subtitle = " ")
}
\arguments{
\item{individual}{Output from \code{rmap_individual}.}

\item{title}{A character string that specifies the title 
of the individualized attribute diagram plot. 
Defaults to \code{"Individualized Attribute Diagram"}.}

\item{subtitle}{A character string that specifies the
subtitle.  Defaults to \code{""}.}
}
\value{
A graphic that draws the attribute diagram
together with a \code{title} and a \code{subtitle}.
}
\description{
An annotated indivualized attribute diagram plot using results from 
\code{rmap_individual}
}
\examples{
data(random_sample_example)
xxx = random_sample_example
e = xxx$e
t = xxx$t
r = xxx$r
t_star = 10
design = "random_sample"
epsilon = length(e)^(-1/3)
N_bootstraps = 100
set.seed(1)
individual = rmap_individual(e, t, r, t_star, design, epsilon, N_bootstraps)
title = "This is to show off the title of the individualized attribute diagram"
subtitle = "I have nothing more to say"
pretty_individual_risk_plot(individual, title, subtitle)

}
