% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_risk_plot.R
\name{pretty_risk_plot}
\alias{pretty_risk_plot}
\title{An annotated attribute diagram plot using \code{rmap_answers}}
\usage{
pretty_risk_plot(rmap_answers, title = "Attribute Diagram")
}
\arguments{
\item{rmap_answers}{Output from \code{rmap} or the
\code{rmap_answers} object from \code{rmap_random_sample}
or \code{rmap_weighted_sample}.}

\item{title}{A character string that specifies the title 
of the attribute diagram plot. Defaults to 
\code{"Attribute Diagram"}.}
}
\value{
A graphic that draws the attribute diagram
together with a \code{title} and a subtitle.
The subtitle contains the goodness of fit test statistic
and its p-value.
}
\description{
An annotated attribute diagram plot using \code{rmap_answers}
}
\examples{
data(random_sample_example)
xxx = random_sample_example
e = xxx$e
t = xxx$t
r = xxx$r
t_star = 10
K = 4
N_bootstraps = 100
set.seed(1)
rmap_answers = rmap_random_sample(e, t, r, t_star, K, N_bootstraps)
title = "This is to show off the title of the attribute diagram plot"
pretty_risk_plot(rmap_answers, title)

}
