% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_roc_plot.R
\name{pretty_roc_plot}
\alias{pretty_roc_plot}
\title{An annotated ROC plot using \code{rmap_answers}}
\usage{
pretty_roc_plot(rmap_answers, title = "ROC Plot")
}
\arguments{
\item{rmap_answers}{Output from \code{rmap} or the
\code{rmap_answers} object from \code{rmap_random_sample}
or \code{rmap_weighted_sample}.}

\item{title}{A character string that specifies the title 
of the ROC plot. Defaults to \code{"ROC Plot"}.}
}
\value{
A graphic that draws the ROC plot
together with a \code{title} and a subtitle.
The subtitle contains the concordance estimate
and confidence interval.
}
\description{
An annotated ROC plot using \code{rmap_answers}
}
\examples{
data(random_sample_example)
xxx = random_sample_example
e = xxx$e
t = xxx$t
r = xxx$r
t_star = 10
K = 4
N_bootstraps = 100
set.seed(1)
rmap_answers = rmap_random_sample(e, t, r, t_star, K, N_bootstraps)
title = "This is to show off the title of the ROC plot"
pretty_roc_plot(rmap_answers, title)

}
