% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapr.R
\docType{data}
\name{random_sample_example}
\alias{random_sample_example}
\title{A random sample example}
\format{A data.frame with 300 rows and 3 variables}
\usage{
data(random_sample_example)
}
\description{
A dataset gotten using a random sample design for trying out \code{rmap}.
 The variables are
}
\details{
\itemize{
 \item e. The event
 \item t. The time of the event.
 \item r. The assigned risk.
}
}
\keyword{rmap_datasets}
