% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_to_df.R
\name{shape_to_df}
\alias{shape_to_df}
\title{shape_to_df}
\usage{
shape_to_df(shape = NULL, shapeColumn = NULL)
}
\arguments{
\item{shape}{Default = NULL. SpatialPolygonDataFrame map (e.g. rmap::mapCountries)}

\item{shapeColumn}{(Optional). Default = NULL. If different from subRegion.}
}
\value{
Fortified ggplot dataframe (e.g. rmap::mapCountriesdf)
}
\description{
Given a SpatialPolygonDataFrame map (e.g. rmap::mapCountries) will choose the
corresponding fortified ggplot dataframe (e.g. rmap::mapCountriesdf)
}
\examples{
library(rmap)
shape = rmap::mapCountries
df = rmap::shape_to_df(shape)
}
\keyword{find}
\keyword{map,}
