% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_shape.R
\name{tidy_shape}
\alias{tidy_shape}
\title{tidy_shape}
\usage{
tidy_shape(shape, shapeColumn = NULL)
}
\arguments{
\item{shape}{Default = NULL. SpatialPolygonDataFrame map (e.g. rmap::mapCountries)}

\item{shapeColumn}{(Optional). Default = NULL. If different from subRegion.}
}
\value{
Fortified ggplot dataframe (e.g. rmap::mapCountriesdf)
}
\description{
Given a SpatialPolygonDataFrame map (e.g. rmap::mapCountries) will convert to a tidy dataframe
}
\examples{
library(rmap)
shape = rmap::mapCountries
df = rmap::tidy_shape(shape)
}
\keyword{find}
\keyword{map,}
