% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapr.R
\docType{data}
\name{two_model_comparison_example}
\alias{two_model_comparison_example}
\title{A random sample with two risk models}
\format{A data.frame with 1000 rows and 4 variables}
\usage{
data(two_model_comparison_example)
}
\description{
A dataset gotten using a random sample design
containing two risk models for trying out \code{rmap}.
 The variables are
}
\details{
\itemize{
 \item e. The event
 \item t. The time of the event.
 \item r1. The risk assigned by model 1.
 \item r2. The risk assigned by model 2, a noisy version of model 1.
}
}
\keyword{rmap_datasets}
