% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapr.R
\docType{data}
\name{two_stage_sample_example}
\alias{two_stage_sample_example}
\title{A two-stage sample example}
\format{A data.frame with 210 rows and 4 variables}
\usage{
data(two_stage_sample_example)
}
\description{
A dataset gotten using a two-stage sample design for trying out \code{rmap}.
This is the second stage sample.  To obtain the counts from the first stage
use \code{N_first_stage}
 The variables are
}
\details{
\itemize{
 \item e. The event
 \item t. The time of the event.
 \item r. The assigned risk.
 \item category. The two-stage sample category. 
}
}
\keyword{rmap_datasets}
