% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_fn.R
\name{weight_fn}
\alias{weight_fn}
\title{Calculate cohort weights based on cohort and target categories}
\usage{
weight_fn(cohort_category, target_category)
}
\arguments{
\item{cohort_category}{A character vector of length equal to
the number of people in the cohort sample describing the
category of each person in the cohort sample.}

\item{target_category}{Another character vector describing the
categories of a target sample from the target population.}
}
\value{
A list containing the elements category weights,
code, message, and weight.
}
\description{
If your cohort sample, broken down into relevant categories,
does not match your target population, use this function to
calculate weights to compensate for the mismatch.
}
\examples{
set.seed(1)
NNN = 100
weighted_example = weighted_example_fn(NNN)
cohort_sample = weighted_example$cohort_sample
target_sample = weighted_example$target_sample
list(tail_cohort_sample = tail(cohort_sample),
     tail_target_sample = tail(target_sample))
cohort_category = cohort_sample$category
target_category = target_sample$category
weight = weight_fn(cohort_category, target_category)
list(weight = weight,
     head_cohort_sample_with_weight_appended = head(cbind(cohort_sample, weight = weight$weight)))

}
