% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapr.R
\docType{data}
\name{weighted_example_cohort_sample}
\alias{weighted_example_cohort_sample}
\title{The cohort sample of a weighted sample example}
\format{A data.frame with 224 rows and 4 variables}
\usage{
data(weighted_example_cohort_sample)
}
\description{
A dataset whose covariates do not match those of the target population
The related target sample is \code{weighted_example_target_sample}
The variables of the cohort sample are
}
\details{
\itemize{
 \item e. The event
 \item t. The time of the event.
 \item r. The assigned risk.
 \item category. The joint-covariate category to be matched with that of the target sample
}
}
\keyword{rmap_datasets}
