% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_example_fn.R
\name{weighted_example_fn}
\alias{weighted_example_fn}
\title{Generate the data required to showcase weighted rmap}
\usage{
weighted_example_fn(NNN)
}
\arguments{
\item{NNN}{A positive integer defining how many people are
generated by the random sample before oversampling and
oversampling.}
}
\value{
A list containing
(a) \code{cohort_sampling_probability_dictionary}, a named vector
showing the sampling probabilities used,
(b) \code{cohort_sample} A data.frame containing columns
\code{x_1}, \code{x_2}, \code{category},
\code{sam} (the sampling probability), \code{r_A}, \code{r_B}, \code{t_0},
\code{t_1}, \code{ttt}, \code{eee},
(c) \code{target_sample} A data.frame containing columns of the same
names as those of \code{cohort_sample} gotten in the same way as
\code{cohort_sample} but with \code{sampling_probability_dictionary}
values all unity,
(d) \code{t_star} The right end point of duration of study,
used to define \code{r_A} and \code{r_B}.
}
\description{
Weighted rmap requires a cohort sample with columns
`\code{e}, \code{t}, \code{r}, \code{category}
as well as a vector of target categories.
This function generates such data.
For the risk vector \code{r}, you may choose between
\code{r_A} equal to the true probability of disease
and \code{r_B} a probability that accounts for
only half the effect of the covariate \code{x_2}
which is one of the two covariates both determining
disease and defining the \code{category} vector.
Cohort and target samples are obtained via calls
to \code{cohort_fn} with different
versions of the parameter \code{sampling_probability_dictionary}
}
\examples{
NNN = 13
weighted_example_fn(NNN)

}
