% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapr.R
\docType{data}
\name{weighted_example_target_sample}
\alias{weighted_example_target_sample}
\title{The target sample of a weighted sample example}
\format{A data.frame with 400 rows and 3 variables}
\usage{
data(weighted_example_target_sample)
}
\description{
A dataset sampled from  the target population
This provides the dataset \code{weighted_example_target_sample}
with joint-covariate category data used by \code{rmap} to obtain
weights to perform a weighted \code{rmap} analysis.  The other variables 
\code{x_1} and \code{x_2} were presumably used to create determine
\code{category}, but you may ignore them.
}
\details{
\itemize{
 \item x_1. A variable you may ignore
 \item x_2. Another variable you may ignore
 \item category. The joint-covariate category to be matched with that of the cohort sample
}
}
\keyword{rmap_datasets}
