% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockDiagFn.R
\name{BlockDiagFn}
\alias{BlockDiagFn}
\title{Compute the block diagonal matrix given a list of matrices}
\usage{
BlockDiagFn(boxes)
}
\arguments{
\item{boxes}{A list of boxes (matrices) to 
combine to make the block diagonal matrix.}
}
\value{
A block diagonal matrix.  Credit to the simex package for this algorithm.
}
\description{
Compute the block diagonal matrix given a list of matrices
}
\examples{
a = matrix(1, 2, 2)
b = matrix(2, 3, 3)
c = matrix(3, 4, 5)
BlockDiagFn(list(a, b, c))

}
