% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_find.R
\name{map_find}
\alias{map_find}
\title{map_find}
\usage{
map_find(data)
}
\arguments{
\item{data}{data to process}
}
\value{
dataframe with modified subRegions, subRegion shapefile and subRegion type
}
\description{
Given a data.frame with a subRegion column, this function searches for an appropriate map from the pre-loaded rmap maps.
}
\examples{
library(rmap)
data = data.frame(subRegion=c("FL","ID","MO"),value=c(-2,3,14))
rmap::map_find(data)
}
\keyword{find}
\keyword{map,}
