% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_find_df.R
\name{map_find_df}
\alias{map_find_df}
\title{map_find_df}
\usage{
map_find_df(data)
}
\arguments{
\item{data}{data table to be processed}
}
\value{
dataframe with modified subRegions, subRegion shapefile and subRegion type
}
\description{
Given a data.frame with a subRegion column, this function searches for an appropriate map from the pre-loaded rmap maps.
}
\keyword{find}
\keyword{map,}
