% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{map_plot}
\usage{
map_plot(
  data = NULL,
  fillColumn = NULL,
  shapeColumn = NULL,
  shape = NULL,
  theme = NULL,
  show = T,
  palette = "Spectral",
  legendType = "kmeans",
  labels = F,
  labelRepel = 0,
  labelColor = "black",
  labelSize = 2,
  labelAlpha = 1,
  labelFill = NA,
  labelBorderSize = NA,
  width = 9,
  height = 7,
  legendShow = F,
  legendTitle = NULL,
  legendBreaksn = 5,
  legendBreaks = NULL,
  pdfpng = "png",
  underLayer = NULL,
  underLayerColor = "gray40",
  underLayerFill = "gray90",
  underLayerLwd = 0.1,
  underLayerAlpha = 1,
  underLayerLabels = F,
  overLayerLabels = F,
  overLayer = NULL,
  overLayerColor = "gray40",
  overLayerFill = NA,
  overLayerLwd = 0.2,
  overLayerAlpha = 0,
  zoom = -1,
  zoomx = NULL,
  zoomy = NULL,
  asp = 1.2,
  save = T,
  fileName = "map",
  folder = paste(getwd(), "/outputs", sep = ""),
  row = NULL,
  col = NULL,
  title = NULL,
  numeric2Cat_list = NULL,
  catParam = NULL,
  legendDigits = NULL,
  legendDigitsOverride = NULL,
  legendSingleColor = "white",
  legendSingleValue = F,
  colorNA = "gray50",
  showNA = F,
  ncol = 3,
  size = 16,
  alpha = 1,
  background = F,
  crop = T,
  crop_to_underLayer = F,
  crop_to_overLayer = F,
  transparent = F
)
}
\arguments{
\item{data}{Default = NULL}

\item{fillColumn}{Default = NULL # Or give column with data}

\item{shapeColumn}{(Optional). Default = NULL. If different from subRegion.}

\item{shape}{Default = NULL, Cusotm shape can be provided as a SpatialPolygonDataFrame with features corresponding to subRegion columns in the data provided.}

\item{theme}{Default = NULL}

\item{show}{Default = T. Print maps in console as they are processed.}

\item{palette}{Default = "Spectral"}

\item{legendType}{Default = "continuous".}

\item{labels}{Default = FALSE}

\item{labelRepel}{Default = 0,}

\item{labelColor}{Default = "black",}

\item{labelSize}{Default = 3}

\item{labelAlpha}{Default = 0.7,}

\item{labelFill}{Default = NA,}

\item{labelBorderSize}{Default = NA}

\item{width}{Default = 9}

\item{height}{Default = 7}

\item{legendShow}{Default = F}

\item{legendTitle}{Default = NULL}

\item{legendBreaksn}{Default = "5"}

\item{legendBreaks}{Default = NULL}

\item{pdfpng}{Default = "png"}

\item{underLayer}{Default = NULL}

\item{underLayerColor}{Default = "gray40"}

\item{underLayerFill}{Default = "gray40"}

\item{underLayerLwd}{Default = 0.5}

\item{underLayerAlpha}{Default = 1}

\item{underLayerLabels}{Default = F}

\item{overLayerLabels}{Default = F}

\item{overLayer}{Default = NULL}

\item{overLayerColor}{Default = "gray40"}

\item{overLayerFill}{Default = NA}

\item{overLayerLwd}{Default = 0.5}

\item{overLayerAlpha}{Default = 0}

\item{zoom}{Default =-1. Zoom into or out of map. Positive values zoom in and negative out.}

\item{zoomx}{Default = NULL. Zoom into or out of map along x. Positive values zoom in and negative out.}

\item{zoomy}{Default = NULL. Zoom into or out of map along y. Positive values zoom in and negative out.}

\item{asp}{Default = 1.2. Aspect ratio of lat and lon.}

\item{save}{Default = T}

\item{fileName}{Default = "map"}

\item{folder}{Default = paste(getwd(),"/outputs",sep Default = "")}

\item{row}{Default=NULL}

\item{col}{Default=NULL}

\item{title}{Default=NULL}

\item{numeric2Cat_list}{Default=NULL}

\item{catParam}{Default=NULL}

\item{legendDigits}{Default = NULL}

\item{legendDigitsOverride}{Default=NULL}

\item{legendSingleColor}{Default="white"}

\item{legendSingleValue}{Default=F. Change to True to get default single value or provide a numeric value.}

\item{colorNA}{Default = "gray50"}

\item{showNA}{Default = T}

\item{ncol}{Default = 3.  Number of columns to wrap maps}

\item{size}{Default = 12. Text size of plots.}

\item{alpha}{Default = 1. Transparency of fill colors.}

\item{background}{Default = F. Add background water color, border and default underlayer map.}

\item{crop}{Default = T. Crop to boundary data.}

\item{crop_to_underLayer}{Default = F. Crop to the underLayer boundary provided.}

\item{crop_to_overLayer}{Default = F. Crop to the overLayer boundary provided.}

\item{transparent}{Default = F. To make map background transparent for maps without backgrounds.}
}
\value{
Returns the formatted data used to produce chart
}
\description{
This function produce different kinds of maps for the rmap package.
Each figure is accompanied with a csv table.
}
\keyword{charts,}
\keyword{diffplots}
