% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasins}
\alias{mapGCAMBasins}
\title{GCAM Basins}
\format{
A SpatialPolygonsDataFrame
}
\source{
Modified versions of shapefiles from \url{https://zenodo.org/record/4688451#.YdMNTmjMJPY}
}
\usage{
mapGCAMBasins
}
\description{
GCAM Basins
}
\examples{
\dontrun{
 library(sp); library(rmap)
 sp::plot(mapGCAMBasins)
 head(mapGCAMBasins@data)
}
}
\keyword{datasets}
