% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMBasinsdf}
\alias{mapGCAMBasinsdf}
\title{GCAM Basins}
\format{
A SpatialPolygonsDataFrame
}
\source{
From JGCRI confluence page. \url{https://confluence.pnnl.gov/confluence/display/JGCRI/GCAM+Shape+Files}
}
\usage{
mapGCAMBasinsdf
}
\description{
GCAM Basins
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapGCAMBasinsdf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapGCAMBasinsdf)
}
}
\keyword{datasets}
