% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapGCAMReg32df}
\alias{mapGCAMReg32df}
\title{GCAM 32 Regions}
\format{
A SpatialPolygonsDataFrame
}
\source{
From JGCRI confluence page. \url{https://confluence.pnnl.gov/confluence/display/JGCRI/GCAM+Shape+Files}
}
\usage{
mapGCAMReg32df
}
\description{
GCAM 32 Regions
}
\examples{
\dontrun{
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapGCAMReg32df, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapGCAMReg32df)
}
}
\keyword{datasets}
