% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapIntersectGCAMBasin32RegUruguay}
\alias{mapIntersectGCAMBasin32RegUruguay}
\title{Intersection of GCAM Basins and GCAM 32 Regions + Uruguay.}
\format{
sf
}
\source{
JGCRI
}
\usage{
mapIntersectGCAMBasin32RegUruguay
}
\description{
Intersection of GCAM Basins and GCAM 32 Regions + Uruguay.
}
\examples{
\dontrun{
library(rmap)
plot(mapIntersectGCAMBasin32RegUruguay[,"subRegion"])
head(mapIntersectGCAMBasin32RegUruguay)
}
}
\keyword{datasets}
