% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_CI_Fn.R
\name{prob_CI_Fn}
\alias{prob_CI_Fn}
\title{Calculate the 95 percent confidence interval of a probability using a logit transformation}
\usage{
prob_CI_Fn(ps, sds)
}
\arguments{
\item{ps}{A numeric vector of probabilities}

\item{sds}{A numeric vector of standard deviations.  The elements
in this vector correspond to the elements in \code{ps} above}
}
\value{
A matrix with two columns \code{lower} and \code{upper}
and a row for each element in \code{ps} and \code{sds}.
}
\description{
Calculate the 95 percent confidence interval of a probability using a logit transformation
}
\examples{
prob_CI_Fn(ps = 0.99, sds = 0.01)
prob_CI_Fn(ps = c(0.98, 0.99), sds = c(0.01, 0.01))
set.seed(1)
someCIs = prob_CI_Fn( ps = (1:99) / 100, sds = runif(99, 0.01, 0.05) )

}
