% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmap.R
\docType{package}
\name{rmap}
\alias{rmap}
\title{rmap}
\description{
`rmap` is a comprehensive mapping package to create easily customizable maps and compare across scenarios, years and data classes. The package allows creation of difference maps as well as customizable legends, color palettes and styles.
}
\details{
\itemize{
\item Github: https://github.com/JGCRI/rmap
\item Webpage: https://jgcri.github.io/rmap/
\item Cheatsheet: https://jgcri.github.io/rmap/cheatsheet.pdf}
}
