% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmap.R
\docType{package}
\name{rmap}
\alias{rmap}
\title{rmap}
\description{
\itemize{
\item Github: https://github.com/JGCRI/rmap
\item Webpage: https://jgcri.github.io/rmap/
\item Cheatsheet: https://github.com/JGCRI/rmap/blob/master/rmapCheatsheet.pdf}
}
